+++ packages/sector/lib/sector.dart
@@ -45,6 +45,10 @@
 /// // └───────┘
 /// ```
 ///
+/// Grids may optionally implement [EfficientIndexGrid] to provide hints to
+/// consumers about how the grid is laid out in memory, which can be used to
+/// optimize traversal algorithms.
+///
 /// ## Custom Implementations
 ///
 /// The default implementation of [Grid] is [ListGrid], which is a dense grid
@@ -66,9 +70,9 @@
 ///   }
 ///   ```
 ///
-/// - Use [RowsBase] and [ColumnsBase] to implement [Rows] and [Columns]:
+/// - Use [RowsMixin] and [ColumnsMixin] to implement [GridAxis]s:
 ///   ```dart
-///   class _Rows<T> extends Iterable<Iterable<T>> with RowsBase<T> { /*...*/ }
+///   class _Rows<T> extends GridAxis<T> with RowsMixin<T> { /*...*/ }
 ///   ```
 ///
 /// After profiling, you may find that the default implementation is sufficient
@@ -80,15 +84,15 @@ import 'dart:typed_data';
 
 import 'package:sector/sector.dart';
 
-export 'src/base/columns.dart' show Columns, ColumnsBase;
+export 'src/base/axis.dart' show ColumnsMixin, GridAxis, RowsMixin;
 export 'src/base/iterator.dart' show GridIterable, GridIterator;
 export 'src/base/layout_hint.dart' show LayoutHint;
-export 'src/base/rows.dart' show Rows, RowsBase;
 export 'src/base/traversal.dart' show Traversal;
-export 'src/grids/grid.dart' show Grid;
+export 'src/grids/grid.dart' show EfficientIndexGrid, Grid;
 export 'src/grids/list.dart' show ListGrid;
 export 'src/traverse/draw_line.dart' show drawLine;
 export 'src/traverse/row_major.dart' show rowMajor;
 export 'src/utils/grid_extension.dart' show GridExtension;
 export 'src/utils/grid_impl.dart' show GridImpl;
 export 'src/utils/octant.dart' show Octant;
+export 'src/views/unmodifiable_grid_view.dart' show UnmodifiableGridView;