+++ packages/sector/lib/src/views/unmodifiable_grid_view.dart
@@ -1,5 +1,5 @@
 import 'package:sector/sector.dart';
-import 'package:sector/src/views/delegating_grid_view.dart.unused';
+import 'package:sector/src/views/delegating_grid_view.dart';
 
 /// A grid view that prevents modification of the underlying grid.
 ///
@@ -21,5 +21,25 @@ final class UnmodifiableGridView<T> extends DelegatingGridView<T> {
   @override
   void clear() {
     throw UnsupportedError('Cannot clear an unmodifiable grid view');
+  }
+
+  @override
+  GridAxis<T> get rows => _UnmodifiableGridAxis(view.rows);
+
+  @override
+  GridAxis<T> get columns => _UnmodifiableGridAxis(view.columns);
+}
+
+final class _UnmodifiableGridAxis<T> extends DelegatingGridAxis<T> {
+  const _UnmodifiableGridAxis(super.view);
+
+  @override
+  void removeAt(int index) {
+    throw UnsupportedError('Cannot modify an unmodifiable grid axis');
+  }
+
+  @override
+  void insertAt(int index, Iterable<T> row) {
+    throw UnsupportedError('Cannot modify an unmodifiable grid axis');
   }
 }