+++ packages/sector/test/grids/naive_grid_test.dart
@@ -1,0 +1,11 @@
+import '../src/naive_grid.dart';
+
+import '../src/suites/mutable_growable.dart';
+import '../src/suites/mutable_in_place.dart';
+import '../src/suites/read.dart';
+
+void main() {
+  runGrowableTestSuite(NaiveListGrid.fromRows);
+  runReadOnlyTestSuite(NaiveListGrid.fromRows);
+  runMutableInPlaceTestSuite(NaiveListGrid.fromRows);
+}