+++ packages/sector/test/traverse/draw_line_test.dart
@@ -1,5 +1,3 @@
-import 'package:sector/sector.dart';
-
 import '../_prelude.dart';
 
 void main() {